[CODE] move platform related utils into a separate module
